fix: rename delete-dialog to DeleteDialog
